deprecate config module
update rn

more nits
