+++ ddtrace/__init__.py
@@ -17,7 +17,7 @@ from ._logger import configure_ddtrace_logger
 # configure ddtrace logger before other modules log
 configure_ddtrace_logger()  # noqa: E402
 
-from .settings import _config as config
+from .settings import _global_config as config
 
 
 # Enable telemetry writer and excepthook as early as possible to ensure we capture any exceptions from initialization
