+++ ddtrace/_trace/sampler.py
@@ -22,7 +22,7 @@ from ..internal.logger import get_logger
 from ..internal.rate_limiter import RateLimiter
 from ..internal.sampling import _get_highest_precedence_rule_matching
 from ..internal.sampling import _set_sampling_tags
-from ..settings import _config as ddconfig
+from ..settings import _global_config as ddconfig
 from .sampling_rule import SamplingRule
