+++ ddtrace/contrib/internal/botocore/patch.py
@@ -33,7 +33,7 @@ from ddtrace.internal.utils import get_argument_value
 from ddtrace.internal.utils.formats import asbool
 from ddtrace.internal.utils.formats import deep_getattr
 from ddtrace.llmobs._integrations import BedrockIntegration
-from ddtrace.settings.config import Config
+from ddtrace.settings._config import Config
 from ddtrace.trace import Pin
 
 from .services.bedrock import patched_bedrock_api_call
