+++ ddtrace/internal/remoteconfig/worker.py
@@ -13,7 +13,7 @@ from ddtrace.internal.remoteconfig.constants import REMOTE_CONFIG_AGENT_ENDPOINT
 from ddtrace.internal.remoteconfig.utils import get_poll_interval_seconds
 from ddtrace.internal.service import ServiceStatus
 from ddtrace.internal.utils.time import StopWatch
-from ddtrace.settings import _config as ddconfig
+from ddtrace.settings import _global_config as ddconfig
 
 
 log = get_logger(__name__)
