+++ releasenotes/notes/munir-deprecate-rc-http-configs-206f6c60ff383c05.yaml
@@ -1,7 +1,7 @@
 ---
 deprecations:
   - |
-    tracing: Deprecates support for the following ``ddtrace.config`` attributes:
+    tracing: Internalizes the ``ddtrace.settings.config`` module and deprecates support for the following ``ddtrace.config`` attributes:
     - http
     - http_server
     - trace_headers
